% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examiner_compare.R
\name{examiner_compare}
\alias{examiner_compare}
\title{examiner_compare}
\usage{
examiner_compare(faculty_nr, module_nr, semester_vector, download, FacData)

  ##Default method

  examiner_compare(faculty_nr, module_nr, semester_vector="all", download=FALSE, FacData=NA)
}
\arguments{
\item{faculty_nr}{A numeric value corresponding to a certain faculty. See the function faculty_data to get all faculty numbers.}

\item{module_nr}{A numeric value corresponding to a certain module. See the function list_modules to get all module numbers.}

\item{semester_vector}{A vector of numeric values containing the semester-numbers. By default all semesters are considered. Use the function semester_data to get the specific values.}

\item{download}{Logical. If TRUE the corresponding data is downloaded and used for computing the examiner mean values. If False the faculty data has to be provided using the function faculty_down in the first place.}

\item{FacData}{FacData A List containing the faculty or module data. Typically produced by the function faculty_down.}
}
\description{
This is a function for computing the mean values for the different examiners for one module over all or some chosen semesters.
}
\examples{
# Compute the mean values for the different examiners for the introductory
# course in statistics at the economic faculty over all semesters.

faculty_data() # the required faculty-number is 12
list_modules(12) # the required module-number is 109

# Computing the mean for the economic faculty without providing any data in advance
examiner_compare(faculty_nr = 12, module_nr = 109, download = TRUE)

# Downloading the data for economic faculty using the faculty_down-function
economic_data <- faculty_down(12)
examiner_compare(faculty_nr = 12, module_nr = 109, download = FALSE, FacData = economic_data)
}
