% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty_down.R
\name{faculty_down}
\alias{faculty_down}
\title{faculty_down}
\usage{
faculty_down(facultyNr)
}
\arguments{
\item{facultyNr}{A numeric value corresponding to a certain faculty. See the function faculty_data to get all faculty-numbers.}
}
\value{
A list containing the data of all modules for a certain faculty. Every list-element contains the data for a single module over all semesters.
}
\description{
This function downloads the data for all modules of a faculty. The data can then be used for further analysis.
}
