% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty_mean.R
\name{faculty_mean}
\alias{faculty_mean}
\title{faculty_mean}
\usage{
faculty_mean(faculty_nr, download, FacData)
   ## Default method:
   module_mean(faculty_nr, download=FALSE, FacData=NA)
}
\arguments{
\item{faculty_nr}{A numeric value corresponding to a certain semester. See the function semester_data to get all semesternumbers.}

\item{download}{Logical. If TRUE the corresponding data is downloaded and used for computing the faculty mean value. If False the faculty data has to be provided using the function faculty_down in the first place.}

\item{FacData}{A List containing the faculty data. Typically produced by the function faculty_down.}
}
\description{
Function for computing the mean value of a faculty.
}
