% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_modules.R
\name{list_modules}
\alias{list_modules}
\title{list_modules}
\usage{
list_modules(faculty)
}
\arguments{
\item{faculty}{A numeric value corresponding to a certain faculty. Use the function faculty_data() to display all faculty numbers.}
}
\value{
A data.frame containing the module-label and module-number.
}
\description{
This function downloads the module-label and module-number for all modules of a faculty.
}
\examples{
# Download the module-information for the economic faculty

list_modules(12)

# Download the module-information for the medical faculty

list_modules(3)
}
