% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{module_data}
\alias{module_data}
\title{module_data}
\usage{
module_data(semester_nr, faculty_nr, module_nr)
}
\arguments{
\item{semester_nr}{A numeric value corresponding to a certain semester. See the function semester_data to get all semesternumbers.}

\item{faculty_nr}{A numeric value corresponding to a certain faculty. See the function faculty_data to get all facultynumbers.}

\item{module_nr}{A numeric value corresponding to a certain module. See the function list_modules to get all modulenumbers.}
}
\description{
The function downloads the data for a specific module given the semester-, faculty- and modulenumber.
}
