% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_mean.R
\name{module_mean}
\alias{module_mean}
\title{module_mean}
\usage{
module_mean(semester_vector, faculty_nr, module_nr)
   ## Default method:
   module_mean(semester_vector = "all", faculty_nr= NA, module_nr=NA)
}
\arguments{
\item{semester_vector}{A vector of numeric values. The default value takes all semesters.}

\item{faculty_nr}{A numeric value corresponding to a certain faculty. See the function faculty_data to get all facultynumbers.}

\item{module_nr}{A numeric value corresponding to a certain module. See the function list_modules to get all modulenumbers.}
}
\description{
Function for computing the mean value of a specific module
}
