% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFS.date_compare.R
\name{plotFS.date_compare}
\alias{plotFS.date_compare}
\title{plotFS.date_compare}
\usage{
\method{plotFS}{date_compare}(x)
}
\arguments{
\item{x}{An object of class date_compare}
}
\description{
plotFS.date_compare
}
