% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFS.examiner_compare.R
\name{plotFS.examiner_compare}
\alias{plotFS.examiner_compare}
\title{plotFS.examiner_compare}
\usage{
\method{plotFS}{examiner_compare}(x)
}
\arguments{
\item{x}{object of class examiner_compare}
}
\description{
plotFS.examiner_compare
}
