% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFS.examiner_copmare.R, R/plotFS.fac_mean.R
\name{plotFS.fac_mean}
\alias{plotFS.fac_mean}
\title{plotFS.examiner_compare}
\usage{
\method{plotFS}{fac_mean}(x)

\method{plotFS}{fac_mean}(x)
}
\arguments{
\item{x}{object of class examiner_compare}

\item{x}{Object of the fac_mean}
}
\description{
plotFS.examiner_compare

plotFS.fac_mean
}
