% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFS.module_mean.R
\name{plotFS.module_mean}
\alias{plotFS.module_mean}
\title{plotFS.module_mean}
\usage{
\method{plotFS}{module_mean}(x)
}
\arguments{
\item{x}{Object of the module_mean}
}
\description{
plotFS.module_mean
}
