% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFS.date_compare.R
\name{print.date_compare}
\alias{print.date_compare}
\title{plotFS.date_compare}
\usage{
\method{print}{date_compare}(obj)
}
\arguments{
\item{x}{An Object of the class date_compare}
}
\description{
plotFS.date_compare
}
