% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_examiner_compare.R
\name{print.examiner_compare}
\alias{print.examiner_compare}
\title{print examiner compare}
\usage{
\method{print}{examiner_compare}(x, ...)
}
\arguments{
\item{x}{eaxaminer_compare object}
}
\description{
print examiner compare
}
