% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_faculty_mean.R
\name{print.fac_mean}
\alias{print.fac_mean}
\title{print faculty mean}
\usage{
\method{print}{fac_mean}(x, ...)
}
\arguments{
\item{x}{faculty mean object}

\item{...}{...}
}
\description{
print faculty mean
}
