% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_module-mean.R
\name{print.module_mean}
\alias{print.module_mean}
\title{module mean}
\usage{
\method{print}{module_mean}(x, ...)
}
\arguments{
\item{x}{module mean object}

\item{...}{...}
}
\description{
module mean
}
