% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semester_data.R
\name{semester_data}
\alias{semester_data}
\title{semester_data}
\usage{
Usage text here
}
\arguments{
\item{semester}{The expression for the semesterterm. The default value is all semesters}
}
\value{
Data.Frame of semester numbers corresponding to the semesterterm
}
\description{
Description text here
}
