% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_compare.R
\name{date_compare}
\alias{date_compare}
\title{date_compare}
\usage{
date_compare(faculty_nr, module_nr, semester_vector, download, FacData)

  ## Default method:

  date_compare(faculty_nr, module_nr, semester_vector="all", download=FALSE, FacData=NA)
}
\arguments{
\item{faculty_nr}{A numeric value corresponding to a certain semester. See the function semester_data to get all semesternumbers.}

\item{semester_vector}{A vector of numeric values containing the semester-numbers. By default all semesters are considered. Use the function semester_data to get the specific values.}

\item{download}{Logical. If TRUE the corresponding data is downloaded and used for computing the module mean values per semester. If False the faculty data has to be provided using the function faculty_down in the first place.}

\item{FacData}{FacData A List containing the faculty or module data. Typically produced by the function faculty_down.}

\item{modul_nr}{A numeric value corresponding to a certain module. See the function list_modules to get all modulenumbers}
}
\value{
The Return is an object of class "date_compare" and contains the information of the number exam date (Date 1, Date 2, etc.) of the chosen module and the mean values of the corresponding dates over the chosen semesters.
}
\description{
This is a function for computing the mean values for each exam date per semester over all or some chosen semesters
}
\examples{
# Compute the mean values for the different exam dates within a semester for the introductory course in statistics at the economic faculty over all semesters.

faculty_data() # the required faculty-number is 12
list_modules(12) # the required module-number is 109

# Computing the mean for the economic faculty without providing any data in advance
date_compare <- function(faculty_nr = 12, module_nr = 109, semester_vector = "all", download = TRUE)

# Downloading the data for economic faculty using the faculty_down-function
economic_data <- faculty_down(12)
date_compare <- function(faculty_nr = 12, module_nr = 109, semester_vector = "all", download = FALSE, FacData = economic_data)
}
