% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty_mean.R
\name{faculty_mean}
\alias{faculty_mean}
\title{faculty_mean}
\usage{
faculty_mean(faculty_nr, download, FacData)

   ## Default method:

   faculty_mean(faculty_nr, download=FALSE, FacData=NA)
}
\arguments{
\item{faculty_nr}{A numeric value corresponding to a certain semester. See the function semester_data to get all semesternumbers.}

\item{download}{Logical. If TRUE the corresponding data is downloaded and used for computing the faculty mean value. If False the faculty data has to be provided using the function faculty_down in the first place.}

\item{FacData}{A List containing the faculty data. Typically produced by the function faculty_down.}
}
\value{
The Return is an object of class "fac_mean" and contains the mean value of the faculty and the faculty-name.
}
\description{
This is a function for computing the mean value of a faculty over all semesters.
}
\examples{
# Computing the mean for the economic faculty without providing any data in advance
faculty_mean(12, download = TRUE)

# Downloading the data for economic faculty using the faculty_down-function
economic_data <- faculty_down(12)
faculty_mean(12, download = FALSE, FacData = economic_data)
}
