% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_modules.R
\name{list_modules}
\alias{list_modules}
\title{list_modules}
\usage{
list_modules(faculty)
}
\arguments{
\item{faculty}{A numeric value corresponding to a certain faculty. Use the function faculty_data() to display all faculty numbers.}
}
\description{
The function downloads the faculty-specific module-information
}
