% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{module_data}
\alias{module_data}
\title{module_data}
\usage{
module_data(semester_nr, faculty_nr, module_nr)
}
\arguments{
\item{semester_nr}{A numeric value corresponding to a certain semester. See the function semester_data to get all semester-numbers.}

\item{faculty_nr}{A numeric value corresponding to a certain faculty. See the function faculty_data to get all faculty-numbers.}

\item{module_nr}{A numeric value corresponding to a certain module. See the function list_modules to get all modulen-umbers.}
}
\value{
The result is a data.frame containing the data for a module for one semester. Depending on the module, it is possible that there are several rows, containing the data for several exam dates.
}
\description{
The function downloads the data for a specific module given the semester-, faculty- and module-number.
}
\examples{
 # Download the data for the introductory course in statistics at the economic faculty for the summer semester 2017.
# First get the information for specific semester-, faculty- and module-number

semester_data() # the required semester-number is 65
faculty_data() # the required faculty-number is 12
list_modules(12) # the required module-number is 109
module_data(65,12,109)
}
