% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.date_compare.R
\name{print.date_compare}
\alias{print.date_compare}
\title{print.date_compare}
\usage{
\method{print}{date_compare}(x, ...)
}
\arguments{
\item{x}{An Object of the class date_compare}

\item{...}{...}
}
\description{
print.date_compare
}
