% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semester_data.R
\name{semester_data}
\alias{semester_data}
\title{semester_data}
\usage{
semester_data(semester)

    ## Default method:

    faculty_data(semester = "all")
}
\arguments{
\item{semester}{A is an object of the class character. If semester = "all" all semester-numbers are displayed. If a single semester is entered, it has to be the same expression that is used on the FlexStat-Platform.}
}
\value{
Data.Frame of semester-numbers corresponding to the semesterterm
}
\description{
This is a function to download and display the semester-number corresponding to a certain semester for the FlexStat-Platform at the Goerg-August University in Goettingen.
}
\details{
FlexStat stores every faculty under a certain value. These values have to be used for all function of the GoeFlexStatCrawler-Package.
Below are faculty-labels and the corresponding faculty-numbers
WS 2003/2004    38

SS 2004         39

...

SS 2019         69

WS 2019/2020    70
}
\examples{
semester_data("all") # Downloading all semester-numbers
semester_data("SS 2004") # Downloading a single semester-number
}
