% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty_down.R
\name{single_request}
\alias{single_request}
\title{faculty_down}
\usage{
faculty_down(facultyNr)
}
\arguments{
\item{facultyNr}{A numeric value corresponding to a certain faculty. See the function faculty_data to get all facultynumbers.}
}
\description{
Downloads the data for all modules of a faculty. The data can than be used for further analysis.
}
