% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty_down.R
\name{single_request}
\alias{single_request}
\title{single_request}
\usage{
single_request(Semester,Fakultaet,Modul)
}
\arguments{
\item{Semester}{A numeric value corresponing to a semester-number.}

\item{Fakultaet}{A numeric value corresponing to a faculty-number.}

\item{Modul}{A numeric value corresponing to a module-number.}
}
\description{
Help-function, needed for the faculty_down_function
}
